--- doc/ref/modules/all/salt.modules.docker.rst
@@ -1,7 +1,0 @@
-===================
-salt.modules.docker
-===================
-
-.. automodule:: salt.modules.docker
-    :members:
-    :exclude-members: cp, freeze, unfreeze
