+++ doc/ref/modules/all/salt.modules.dockermod.rst
@@ -1,0 +1,7 @@
+======================
+salt.modules.dockermod
+======================
+
+.. automodule:: salt.modules.dockermod
+    :members:
+    :exclude-members: cp, freeze, unfreeze
